Removing unecessary __init__ and moving testing utils
